backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
pydantic>=2.7.1
python-dotenv>=1.0.1
httpx>=0.27.0
groq>=1.1.1
redis>=5.0.4
pytest>=8.2.0
pytest-asyncio>=0.23.6
